@@ -11,18 +11,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.1",
-    "@angular/common": "^4.0.1",
-    "@angular/compiler": "^4.0.1",
-    "@angular/compiler-cli": "^4.0.1",
-    "@angular/core": "^4.0.1",
-    "@angular/forms": "^4.0.1",
-    "@angular/http": "^4.0.1",
-    "@angular/platform-browser": "^4.0.1",
-    "@angular/platform-browser-dynamic": "^4.0.1",
-    "@angular/platform-server": "^4.0.1",
-    "@angular/router": "^4.0.1",
+    "@angular/animations": "^5.0.0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/platform-server": "^5.0.0",
+    "@angular/router": "^5.0.0",
     "@ng-bootstrap/ng-bootstrap": "^1.0.0-alpha.27",
+    "@types/clockpicker": "0.0.0",
     "angular-calendar": "^0.19.0",
     "angular-confirmation-popover": "^3.1.1",
     "angular-date-value-accessor": "0.0.2",
@@ -30,38 +30,41 @@
     "animate.css": "3.1.1",
     "async": "^2.5.0",
     "bootstrap": "^3.3.7",
-    "chart.js": "^2.6.0",
+    "chart.js": "^2.7.1",
+    "clockpicker": "0.0.7",
     "core-js": "^2.4.1",
     "csv-parser": "^1.11.0",
-    "d3": "^4.9.1",
-    "date-fns": "^1.28.5",
+    "d3": "^4.11.0",
+    "date-fns": "^1.29.0",
     "dateformat": "^2.0.0",
-    "express": "^4.15.3",
+    "express": "^4.16.2",
     "font-awesome": "^4.7.0",
     "jquery": "^3.1.0",
     "jquery-slimscroll": "^1.3.8",
     "jquery-sparkline": "^2.4.0",
+    "jsonwebtoken": "^8.1.0",
     "jvectormap": "1.2.2",
-    "method-override": "^2.3.9",
-    "metismenu": "^2.5.0",
-    "mongoose": "^4.11.5",
-    "morgan": "^1.8.2",
+    "method-override": "^2.3.10",
+    "metismenu": "^2.7.1",
+    "mongoose": "^4.13.0",
+    "morgan": "^1.9.0",
     "muler": "^1.1.1",
-    "ng-select": "^1.0.0-beta.6",
+    "ng-select": "^1.0.0-rc.3",
     "ng2-charts": "^1.6.0",
     "ng2-toastr": "^4.1.2",
     "ng2-validation": "^4.2.0",
     "ngx-bootstrap": "^1.6.6",
-    "ngx-dropzone-wrapper": "^4.6.2",
-    "ngx-pagination": "^3.0.1",
+    "ngx-dropzone-wrapper": "^4.6.6",
+    "ngx-pagination": "^3.0.3",
     "peity": "^3.2.1",
-    "rxjs": "^5.1.0"
+    "rxjs": "^5.5.2",
+    "ts-md5": "^1.2.2"
   },
   "devDependencies": {
-    "@angular/cli": "^1.0.0",
-    "@angular/compiler-cli": "^2.4.0",
+    "@angular/cli": "^1.5.0",
+    "@angular/compiler-cli": "^5.0.0",
     "@types/jasmine": "2.5.38",
-    "@types/node": "~6.0.60",
+    "@types/node": "^6.0.90",
     "codelyzer": "~2.0.0",
     "jasmine-core": "~2.5.2",
     "jasmine-spec-reporter": "~3.2.0",
@@ -74,7 +77,7 @@
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
-    "typescript": "~2.3",
-    "zone.js": "^0.8.16"
+    "typescript": "^2.4.2",
+    "zone.js": "^0.8.18"
   }
 }
