@@ -10,11 +10,12 @@ SchemaFactory.init();
 var ServicesFactory = rootRequire('app/utils/ServicesFactory');
 var ORFactory = rootRequire('app/utils/ORFactory');
 var ScenarioFactory = rootRequire('app/utils/ScenarioFactory');
+var AuthUtil = rootRequire('app/utils/AuthUtil');
 
 var allowCrossDomain = function(req, res, next) {
     res.header('Access-Control-Allow-Origin', "*");
     res.header('Access-Control-Allow-Methods', 'GET,PUT,POST,DELETE');
-    res.header('Access-Control-Allow-Headers', 'Content-Type,Cache-Control,Content-Disposition,X-Requested-With');
+    res.header('Access-Control-Allow-Headers', 'Content-Type,Cache-Control,Content-Disposition,X-Requested-With,authorization');
     next();
 }
 
@@ -33,6 +34,7 @@ var clearAllOutPutFile = function(){
 
 module.exports = function(app) {
     app.use(allowCrossDomain);
+    AuthUtil.apiInit(app);
     ServicesFactory.apiInit(app);
     ORFactory.apiInit(app);
     ScenarioFactory.apiInit(app);
