@@ -1,11 +1,12 @@
 module.exports = {
-    dbURL: "mongodb://10.37.129.2:27017/projectManagement",
-    //dbURL: "mongodb://localhost:27017/projectManagement",
+    // dbURL: "mongodb://10.37.129.2:27017/projectManagement",
+    dbURL: "mongodb://localhost:27017/projectManagement",
     dataPath : "src/",
     outputPath : "src/assets/or/output/",
     orInputPath : "src/assets/or/temp/",
     systemPath: "Z:/Users/Alex.Chen/git/project-management-windows/src/assets/or/",
     //systemPath: "src/assets/or/",
     uploadPath: "src/assets/or/upload/",
-    command: "ProductionPlan"
+    command: "ProductionPlan",
+    secret: "bigbound"
 }
