@@ -1,3 +1,3 @@
 // export const HTTP_BASE = "http://10.211.55.4:8001";
-export const HTTP_BASE = "http://10.211.55.6:8001";
-//export const HTTP_BASE = "http://localhost:8001";
+// export const HTTP_BASE = "http://10.211.55.6:8001";
+export const HTTP_BASE = "http://localhost:8001";
