@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
-import { Http, Response } from '@angular/http';
+import { Response } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 import { HTTP_BASE } from 'app/config';
 import { GeneralService } from 'app/services/general.service';
@@ -11,7 +12,7 @@ import 'rxjs/add/observable/forkJoin'
 @Injectable()
 export class OrResultService extends GeneralService {
 
-  constructor(http: Http) { super(http); }
+  constructor(http: HttpClient) { super(http); }
   
     public find(conditions: any){
       return super.find("orResult", conditions, undefined);
