@@ -1,11 +1,12 @@
 import { Component, OnInit, ViewChild} from '@angular/core';
 import { ItemService } from 'app/services/item.service';
+import { AuthenticationService } from 'app/utils/authentication.service';
 
 @Component({
   selector: 'app-wizard',
   templateUrl: './wizard.component.html',
   styleUrls: ['./wizard.component.css'],
-  providers: [ ItemService]
+  providers: [ ItemService, AuthenticationService]
 })
 export class WizardComponent implements OnInit {
   @ViewChild('formWizard') public formWizard;
